+++ src/plone/restapi/services/content/add.py
@@ -8,8 +8,6 @@ from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.services import Service
 from plone.restapi.services.content.utils import add
 from plone.restapi.services.content.utils import create
-from plone.restapi.services.content.utils import rename
-from Products.Archetypes.interfaces import IBaseObject
 from Products.CMFPlone.utils import safe_hasattr
 from zExceptions import BadRequest
 from zExceptions import Unauthorized
@@ -79,14 +77,10 @@ class FolderPost(Service):
         if temporarily_wrapped:
             obj = aq_base(obj)
 
-        if not IBaseObject.providedBy(obj):
+        if not getattr(deserializer, 'notifies_create', False):
             notify(ObjectCreatedEvent(obj))
 
-        obj = add(self.context, obj)
-
-        # Rename if generated id
-        if not id_:
-            rename(obj)
+        obj = add(self.context, obj, rename=not bool(id_))
 
         self.request.response.setStatus(201)
         self.request.response.setHeader('Location', obj.absolute_url())